requirements.txt
Flask==1.1.2
google-cloud-datastore==2.1.0
google-cloud-storage==1.37.0
gunicorn==20.0.4
pytz==2020.1
Werkzeug==1.0.1